requirements.txt
Pillow==10.2.0
pandas==2.2.1
openpyxl==3.1.2 